# [Enhancement] Record group lowerBoundCost to prune plan