… fe/fe-core/src/test/resources/sql/query_dump/deep_join_cost.json