## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

In SQL like:
```
select * from A join B
                         join C
                         join D
                         join E
                         join F
                         join .....
```
34 inner join

In memo
```
Best Plan
level5:              ABCDEF
                    /      \
level4:           ABCDE     F
                 /     \
level3:         ABCD    E
               /    \
level3:      ABC     D
            /   \
level2:    AB    C
          /  \
level1:  A   B


Bad PlanB:
level5:              ABCDEF
                    /      \
level4:           BCDEF     A
                 /     \
level3:         BCDE    F
               /    \
level3:      BCD     E
            /   \
level2:    BC    D
          /  \
level1:  B    C
```

In theory, we found BC's cost is large than Best PlanA's cost, sr will prune PlanB immediately. But in fact, sr only prune BC[Any(from parent required)] group, and need to compute cost BC[SHUFFLE(from parent required)] group, the cost of BC is relatively large, no matter the propery of enfore is useless, and the process will happend in group BCD,BCDE,BCDEF,  it's will generate a lot of useless but repeat-calculated tasks


I  add a lowerCostBound in group, If the required costBounder is smaller than this, it means that the group will still be drop in the end, and it is meaningless to calculate the cost

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

left outer join + only left-deep tree

before
```
  901ms|-- Optimizer[1] 22s833ms
 1310ms|    -- RuleBaseOptimize[1] 99ms
 1414ms|    -- CostBaseOptimize[1] 22s309ms
 1439ms|        -- OptimizeGroupTask[3474259] 507ms
 1439ms|        -- OptimizeExpressionTask[4082170] 1s422ms
 1439ms|        -- ExploreGroupTask[8163526] 262ms
 1439ms|        -- DeriveStatsTask[4082170] 148ms
 1439ms|        -- ApplyRuleTask[828] 13ms
 1462ms|        -- EnforceAndCostTask[7555990] 17s368ms
23735ms|-- Builder[1] 34ms
Tracer Cost: 1478695us
```

after:
```
 743ms|-- Optimizer[1] 367ms
 767ms|    -- RuleBaseOptimize[1] 224ms
 996ms|    -- CostBaseOptimize[1] 87ms
1016ms|        -- OptimizeGroupTask[569] 0
1016ms|        -- OptimizeExpressionTask[847] 3ms
1017ms|        -- ExploreGroupTask[943] 0
1017ms|        -- DeriveStatsTask[847] 11ms
1020ms|        -- ApplyRuleTask[321] 9ms
1032ms|        -- EnforceAndCostTask[1129] 39ms
1085ms|    -- PhysicalRewrite[1] 19ms
1105ms|    -- PlanValidate[1] 6ms
1111ms|-- Builder[1] 55ms
Tracer Cost: 4323us
```

And this pr depend on https://github.com/StarRocks/starrocks/pull/15358, to avoid statistics update

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5